refactor: entry, ticketing package 분리
