--- backend/src/main/java/com/festago/domain/EntryCodeExtractor.java
@@ -1,6 +1,0 @@
-package com.festago.domain;
-
-public interface EntryCodeExtractor {
-
-    EntryCodePayload extract(String code);
-}
