+++ backend/src/main/java/com/festago/ticket/domain/Ticket.java
@@ -1,10 +1,10 @@
 package com.festago.ticket.domain;
 
 import com.festago.domain.BaseTimeEntity;
-import com.festago.domain.MemberTicket;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.stage.domain.Stage;
+import com.festago.ticketing.domain.MemberTicket;
 import com.festago.zmember.domain.Member;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Entity;
