+++ backend/src/main/java/com/festago/ticketing/application/MemberTicketService.java
@@ -1,16 +1,16 @@
-package com.festago.application;
+package com.festago.ticketing.application;
 
 import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toList;
 
-import com.festago.domain.MemberTicket;
-import com.festago.domain.MemberTicketRepository;
-import com.festago.dto.MemberTicketResponse;
-import com.festago.dto.MemberTicketsResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
+import com.festago.ticketing.domain.MemberTicket;
+import com.festago.ticketing.dto.MemberTicketResponse;
+import com.festago.ticketing.dto.MemberTicketsResponse;
+import com.festago.ticketing.repository.MemberTicketRepository;
 import com.festago.zmember.repository.MemberRepository;
 import java.time.Clock;
 import java.time.Duration;
