+++ backend/src/main/java/com/festago/ticketing/application/TicketingService.java
@@ -1,9 +1,5 @@
-package com.festago.application;
+package com.festago.ticketing.application;
 
-import com.festago.domain.MemberTicket;
-import com.festago.domain.MemberTicketRepository;
-import com.festago.dto.TicketingRequest;
-import com.festago.dto.TicketingResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.ErrorCode;
 import com.festago.exception.NotFoundException;
@@ -11,6 +7,10 @@ import com.festago.ticket.domain.Ticket;
 import com.festago.ticket.domain.TicketAmount;
 import com.festago.ticket.repository.TicketAmountRepository;
 import com.festago.ticket.repository.TicketRepository;
+import com.festago.ticketing.domain.MemberTicket;
+import com.festago.ticketing.dto.TicketingRequest;
+import com.festago.ticketing.dto.TicketingResponse;
+import com.festago.ticketing.repository.MemberTicketRepository;
 import com.festago.zmember.domain.Member;
 import com.festago.zmember.repository.MemberRepository;
 import java.time.Clock;
