+++ backend/src/main/java/com/festago/ticketing/domain/EntryState.java
@@ -1,4 +1,4 @@
-package com.festago.domain;
+package com.festago.ticketing.domain;
 
 import com.festago.exception.ErrorCode;
 import com.festago.exception.InternalServerException;
