+++ backend/src/main/java/com/festago/ticketing/dto/TicketingRequest.java
@@ -1,4 +1,4 @@
-package com.festago.dto;
+package com.festago.ticketing.dto;
 
 import jakarta.validation.constraints.NotNull;
