+++ backend/src/main/java/com/festago/ticketing/dto/TicketingResponse.java
@@ -1,6 +1,6 @@
-package com.festago.dto;
+package com.festago.ticketing.dto;
 
-import com.festago.domain.MemberTicket;
+import com.festago.ticketing.domain.MemberTicket;
 import java.time.LocalDateTime;
 
 public record TicketingResponse(
