+++ backend/src/main/java/com/festago/ticketing/repository/MemberTicketRepository.java
@@ -1,6 +1,7 @@
-package com.festago.domain;
+package com.festago.ticketing.repository;
 
 import com.festago.stage.domain.Stage;
+import com.festago.ticketing.domain.MemberTicket;
 import com.festago.zmember.domain.Member;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
