+++ backend/src/main/java/com/festago/zentry/application/EntryCodeExtractor.java
@@ -1,0 +1,8 @@
+package com.festago.zentry.application;
+
+import com.festago.zentry.domain.EntryCodePayload;
+
+public interface EntryCodeExtractor {
+
+    EntryCodePayload extract(String code);
+}
