+++ backend/src/main/java/com/festago/zentry/application/EntryCodeProvider.java
@@ -1,5 +1,6 @@
-package com.festago.domain;
+package com.festago.zentry.application;
 
+import com.festago.zentry.domain.EntryCodePayload;
 import java.util.Date;
 
 public interface EntryCodeProvider {
