+++ backend/src/main/java/com/festago/zentry/dto/TicketValidationResponse.java
@@ -1,7 +1,7 @@
-package com.festago.dto;
+package com.festago.zentry.dto;
 
-import com.festago.domain.EntryState;
-import com.festago.domain.MemberTicket;
+import com.festago.ticketing.domain.EntryState;
+import com.festago.ticketing.domain.MemberTicket;
 
 public record TicketValidationResponse(
     EntryState updatedState) {
