+++ backend/src/test/java/com/festago/application/EntryServiceTest.java
@@ -7,14 +7,6 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.BDDMockito.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.festago.domain.EntryCode;
-import com.festago.domain.EntryCodePayload;
-import com.festago.domain.EntryState;
-import com.festago.domain.MemberTicket;
-import com.festago.domain.MemberTicketRepository;
-import com.festago.dto.EntryCodeResponse;
-import com.festago.dto.TicketValidationRequest;
-import com.festago.dto.TicketValidationResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.NotFoundException;
 import com.festago.stage.domain.Stage;
@@ -23,6 +15,16 @@ import com.festago.support.MemberFixture;
 import com.festago.support.MemberTicketFixture;
 import com.festago.support.StageFixture;
 import com.festago.support.TimeInstantProvider;
+import com.festago.ticketing.domain.EntryState;
+import com.festago.ticketing.domain.MemberTicket;
+import com.festago.ticketing.repository.MemberTicketRepository;
+import com.festago.zentry.application.EntryCodeManager;
+import com.festago.zentry.application.EntryService;
+import com.festago.zentry.domain.EntryCode;
+import com.festago.zentry.domain.EntryCodePayload;
+import com.festago.zentry.dto.EntryCodeResponse;
+import com.festago.zentry.dto.TicketValidationRequest;
+import com.festago.zentry.dto.TicketValidationResponse;
 import com.festago.zfestival.domain.Festival;
 import com.festago.zmember.domain.Member;
 import java.time.Clock;
