+++ backend/src/test/java/com/festago/application/MemberTicketServiceTest.java
@@ -7,16 +7,17 @@ import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.anyLong;
 import static org.mockito.BDDMockito.given;
 
-import com.festago.domain.MemberTicket;
-import com.festago.domain.MemberTicketRepository;
-import com.festago.dto.MemberTicketResponse;
-import com.festago.dto.MemberTicketsResponse;
 import com.festago.exception.BadRequestException;
 import com.festago.exception.NotFoundException;
 import com.festago.stage.domain.Stage;
 import com.festago.support.MemberFixture;
 import com.festago.support.MemberTicketFixture;
 import com.festago.support.StageFixture;
+import com.festago.ticketing.application.MemberTicketService;
+import com.festago.ticketing.domain.MemberTicket;
+import com.festago.ticketing.dto.MemberTicketResponse;
+import com.festago.ticketing.dto.MemberTicketsResponse;
+import com.festago.ticketing.repository.MemberTicketRepository;
 import com.festago.zmember.domain.Member;
 import com.festago.zmember.repository.MemberRepository;
 import java.time.Clock;
