+++ backend/src/test/java/com/festago/presentation/MemberTicketControllerTest.java
@@ -13,19 +13,19 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.festago.application.EntryService;
-import com.festago.application.MemberTicketService;
-import com.festago.application.TicketingService;
-import com.festago.domain.EntryState;
-import com.festago.dto.EntryCodeResponse;
-import com.festago.dto.MemberTicketFestivalResponse;
-import com.festago.dto.MemberTicketResponse;
-import com.festago.dto.MemberTicketsResponse;
-import com.festago.dto.TicketingRequest;
-import com.festago.dto.TicketingResponse;
 import com.festago.stage.dto.StageResponse;
 import com.festago.support.CustomWebMvcTest;
 import com.festago.support.WithMockAuth;
+import com.festago.ticketing.application.MemberTicketService;
+import com.festago.ticketing.application.TicketingService;
+import com.festago.ticketing.domain.EntryState;
+import com.festago.ticketing.dto.MemberTicketFestivalResponse;
+import com.festago.ticketing.dto.MemberTicketResponse;
+import com.festago.ticketing.dto.MemberTicketsResponse;
+import com.festago.ticketing.dto.TicketingRequest;
+import com.festago.ticketing.dto.TicketingResponse;
+import com.festago.zentry.application.EntryService;
+import com.festago.zentry.dto.EntryCodeResponse;
 import java.nio.charset.StandardCharsets;
 import java.time.LocalDateTime;
 import java.util.stream.LongStream;
